+++ src/draw.ts
@@ -1,10 +1,12 @@
 import * as vscode from 'vscode';
-import path = require('path');
-import fs = require('fs');
-import cheerio = require("cheerio");
+import * as path from 'path';
+import * as fs from 'fs';
+import * as cheerio from "cheerio";
 
 import * as langs from './langs';
 
+// @ts-ignore  
+import html from '../webview/index.html';
 
 export class Draw {
     /** singleton of the currently open panel */
@@ -33,33 +35,25 @@ export class Draw {
 
     private updateCheckStrings = ['', ''];
 
-    /** the dom, currently provided by cheerio */
-    private $: any;
     /** a pseudo-randomly generated value */
     private nonce = nonce();
 
     /**
      * create a new panel, or show the existing one, if available
-     * @param context the current extention context
+     * @param context the current extension context
      */
     public static createOrShow(context: vscode.ExtensionContext): void {
 
         // If we already have a panel, show it.
         if (Draw.currentPanel) {
-            const column = vscode.window?.activeTextEditor?.viewColumn;
-            Draw.currentPanel._panel.reveal(column);
+            Draw.currentPanel._panel.reveal(vscode.window?.activeTextEditor?.viewColumn);
             return;
         }
 
         // Otherwise, create a new panel.
-        const panel = vscode.window.createWebviewPanel(
-            Draw.viewType,
-            'Draw',
-            vscode.ViewColumn.Three,
-            getWebviewOptions(context.extensionUri),
-        );
-
-        Draw.currentPanel = new Draw(panel, context);
+        Draw.currentPanel = new Draw(vscode.window.createWebviewPanel(
+            Draw.viewType, 'Draw', vscode.ViewColumn.Three,
+            getWebviewOptions(context.extensionUri)), context);
     }
 
     private constructor(panel: vscode.WebviewPanel, context: vscode.ExtensionContext) {
@@ -67,35 +61,34 @@ export class Draw {
         this._extensionUri = context.extensionUri;
 
         // Set the webview's initial html content
-        this._update();
+        this._panel.webview.html = html;
 
         // Listen for when the panel is disposed
         // This happens when the user closes the panel or when the panel is closed programmatically
         this._panel.onDidDispose(() => this.dispose(), null, this._disposables);
 
-        this._panel.webview.onDidReceiveMessage(
-            message => {
-                switch (message.command) {
-                    case 'requestCurrentLine':
-                        this.pushCurrentLine();
-                        return;
-                    case 'requestCustom':
-                        this.pushCustom(context);
-                        return;
-                    case 'editCurrentLine':
-                        this.setEditorText(message.text, message.control);
-                        break;
-                    case 'recognize':
-                        context.secrets.get("token").then((token: any) => {
-                            Draw.currentPanel?._panel.webview.postMessage({
-                                command: 'recognize',
-                                token: token,
-                                provider: message.provider
-                            });
+        this._panel.webview.onDidReceiveMessage(message => {
+            switch (message.command) {
+                case 'requestCurrentLine':
+                    this.pushCurrentLine();
+                    return;
+                case 'requestCustom':
+                    this.pushCustom(context);
+                    return;
+                case 'editCurrentLine':
+                    this.setEditorText(message.text, message.control);
+                    break;
+                case 'recognize':
+                    context.secrets.get("token").then((token: any) => {
+                        Draw.currentPanel?._panel.webview.postMessage({
+                            command: 'recognize',
+                            token: token,
+                            provider: message.provider
                         });
-                        return;
-                }
-            },
+                    });
+                    return;
+            }
+        },
             null,
             this._disposables
         );
@@ -117,60 +110,12 @@ export class Draw {
     }
 
     /**
-     * update the webview contents
-     */
-    private _update() {
-        // get the webview index
-        const html = path.resolve(this._extensionUri.fsPath, 'webview', 'index.html');
-
-        // load it into the fake dom
-        this.$ = cheerio.load(fs.readFileSync(html, { encoding: 'utf8' }));
-
-        // add a CSP with the nonce
-        this.$("head").append(`<meta http-equiv="Content-Security-Policy" content="script-src 'unsafe-eval' 'nonce-${this.nonce}';">`);
-
-        // inject dependencies
-        this.inject("node_modules", "@fortawesome", "fontawesome-free", "js", "all.min.js");
-        this.inject("node_modules", "@vscode", "webview-ui-toolkit", "dist", "toolkit.js");
-        this.inject("node_modules", "iink-js", "dist", "iink.min.js");
-        this.inject("webview");
-
-        this._panel.webview.html = this.$.root().html();
-    }
-
-    /**
      * recreate the panel
      * @param panel
      * @param context 
      */
     public static revive(panel: vscode.WebviewPanel, context: vscode.ExtensionContext) {
         Draw.currentPanel = new Draw(panel, context);
-    }
-
-    /**
-     * inject the contents at filepath into the dom
-     * @param filepath 
-     */
-    private inject(...filepath: string[]): void {
-        const absPath = path.resolve(this._extensionUri.fsPath, ...filepath);
-        if (fs.lstatSync(absPath).isDirectory()) {
-            [...fs.readdirSync(absPath)].sort().forEach(p => this.inject(path.join(...filepath, p)));
-            return;
-        }
-
-        const toolkitUri = this._panel.webview.asWebviewUri(vscode.Uri.joinPath(this._extensionUri, ...filepath));
-        let attr = "";
-        switch (path.extname(path.join(...filepath)).substring(1)) {
-            case 'css':
-                this.$("head").append(`<link rel="stylesheet" nonce="${this.nonce}" href="${toolkitUri}">`);
-                break;
-            case 'js':
-                if (filepath.includes("node_modules")) {
-                    attr = `type="module"`;
-                }
-                this.$("body").append(`<script ${attr} nonce="${this.nonce}" src="${toolkitUri}"></script>`);
-                break;
-        }
     }
 
     /**
@@ -330,7 +275,7 @@ export class Draw {
 export function getWebviewOptions(extensionUri: vscode.Uri): vscode.WebviewOptions {
     return {
         enableScripts: true,
-        localResourceRoots: ['webview', 'node_modules'].map((i) => vscode.Uri.joinPath(extensionUri, i))
+        localResourceRoots: []
     };
 }